+++ app/src/main/res/layout/old_files_list_activity.xml
@@ -1,19 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/file_lists_activity_root_layout"
+    android:id="@+id/regaplay_lists_activity_root_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
     <android.support.design.widget.AppBarLayout
-        android:id="@+id/file_lists_activity_appbar"
+        android:id="@+id/regaplay_lists_activity_appbar"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:elevation="6dp"
         android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar">
 
         <android.support.v7.widget.Toolbar
-            android:id="@+id/file_lists_activity_toolbar"
+            android:id="@+id/regaplay_lists_activity_toolbar"
             android:layout_width="match_parent"
             android:layout_height="?attr/actionBarSize"
             android:background="?attr/colorPrimary"
@@ -33,16 +33,20 @@
 
     </android.support.design.widget.AppBarLayout>
 
-        <ListView
-            android:layout_width="match_parent"
+
+    <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_alignParentBottom="true" >
+
+        <ListView android:id="@+id/fileList" android:layout_width="wrap_content"
             android:layout_height="match_parent"
-            android:id="@+id/fileListView"
+            android:foregroundGravity="bottom"
             android:layout_alignParentTop="true"
             android:layout_alignParentStart="true"
-            android:layout_marginTop="58dp"
-            android:translationZ="-1dp"
-            android:layout_alignParentBottom="true" />
-
+            android:layout_marginTop="60dp"
+            />
+    </RelativeLayout>
 
 
 </android.support.design.widget.CoordinatorLayout>
